event planner/event_planner_agent/requirements.txt
Flask==2.3.3
pandas==2.0.3
requests==2.31.0
geopy==2.3.0
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
smtplib3==0.1.0
reportlab==4.0.4
fpdf==1.7.2
pytest==7.4.0
pytest-cov==4.1.0
